pyproject.toml
[tool.poetry]
name = "workflowdemo"
version = "0.1.0"
description = ""
authors = ["pengfei <liu.pengfei@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.21.27"
pandas-profiling = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
